--- vlab implementation/js/webworker.js
@@ -1,45 +1,0 @@
-// webworker.js
-
-// Setup your project to serve `py-worker.js`. You should also serve
-// `pyodide.js`, and all its associated `.asm.js`, `.data`, `.json`,
-// and `.wasm` files as well:
-importScripts("https://cdn.jsdelivr.net/pyodide/v0.21.3/full/pyodide.js");
-
-
-async function loadPyodideAndPackages() {
-  self.result = '';
-  self.error = '';
-  self.pyodide = await loadPyodide({stdout: out => self.result += out + '\n', stderr: err => self.error += err + '\n'});
-  self.postMessage({ loaded: true });
-}
-let pyodideReadyPromise = loadPyodideAndPackages();
-
-self.onmessage = async (event) => {
-  // make sure loading is done
-  await pyodideReadyPromise;
-
-  if (event.data.cmd === "setInterruptBuffer") {
-    pyodide.setInterruptBuffer(event.data.interruptBuffer);
-    return;
-  }
-
-  // Don't bother yet with this line, suppose our API is built in such a way:
-  const { id, python, ...context } = event.data;
-  // The worker copies the context in its own "memory" (an object mapping name to values)
-  for (const key of Object.keys(context)) {
-    self[key] = context[key];
-  }
-  // Now is the easy part, the one that is similar to working in the main thread:
-  try {
-    console.log(python)
-    await self.pyodide.loadPackagesFromImports(python);
-    await self.pyodide.runPythonAsync(python);
-    // console.log(self.result, self.error);
-    self.postMessage({ result: self.result, error: self.error , id });
-  } catch (error) {
-    self.postMessage({ error: error.message, id });
-  }finally {
-    self.result = '';
-    self.error = '';
-  }
-};
